Create calibration_utils.py
I am so grateful for this package and think it is really the most comprehensive calibration package I have come across. One way I thought I could contribute was by improving the quality of the graphical output. There was already a great framework for building a cleaner, more flexible output for those that might wish to take advantage of some of the existing matplotlib params available in scatter, errorbar, and histogram that were presently pretty limited. Further, and importantly, because of the brilliant binning innovation that enabled better control it made sense to provide an outline to the bins for this histogram to better visualize these differences. I hope that you find this contribution helpful and of course feel free to make any further suggestions. E.g. integration of updated calibrated distribution into the histogram plot to visualize this change might be worth considering.